+++ components/display/direct-message/direct-input.tsx
@@ -96,7 +96,7 @@ export const DirectMessageInput = ({
                   </button>
                   <Input
                     id="input"
-                    disabled={isLoading}
+                    // disabled={isLoading}
                     className="px-14 py-6 bg-zinc-200/90 dark:bg-zinc-700/75 border-none border-0 focus-visible:ring-0 focus-visible:ring-offset-0 text-zinc-600 dark:text-zinc-200 overflow-y-scroll"
                     placeholder={`Message ${channelName }`}
                     {...field}